+++ src/app/models/mega_merge/meta_repository/git_sub_module.rb
@@ -25,7 +25,7 @@ module MegaMerge
       attr_accessor :path, :source_branch, :repository, :parent_repository
 
 
-      def config_file
+      def config_files
         path
       end